fix AttributeError

+++ sfa/openstack/osaggregate.py
@@ -259,7 +259,7 @@ class OSAggregate:
 
 
     def delete_instances(self, project_name):
-        project = self.shell.project_get(project_name)
+        project = self.driver.shell.project_get(project_name)
         if not project:
             return 1
         instances = self.driver.shell.db.instance_get_all_by_project(project_name)
@@ -267,7 +267,9 @@ class OSAggregate:
         for instance in instances:
             # deleate this instance's security groups
             for security_group in instance.security_groups:
-                security_group_manager.delete_security_group(security_group.name)
+                # dont delete the default security group
+                if security_group.name != 'default': 
+                    security_group_manager.delete_security_group(security_group.name)
             # destroy instance
             self.driver.shell.db.instance_destroy(instance.instance_id)
         return 1
